requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.12
alembic==1.15.2
pydantic==2.11.4
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
bcrypt==4.0.1
email-validator==2.2.0
